handle empty values in insert into
