+++ crates/datasources/src/common/util.rs
@@ -3,7 +3,7 @@ use std::{fmt::Write, sync::Arc};
 use chrono::{Duration, TimeZone, Utc};
 use datafusion::{
     arrow::{
-        array::{Array, ArrayRef},
+        array::{Array, ArrayRef, UInt64Array},
         compute::{cast_with_options, CastOptions},
         datatypes::{DataType, Field, Schema, TimeUnit},
         error::ArrowError,
@@ -160,6 +160,24 @@ pub fn normalize_batch(batch: &RecordBatch) -> Result<RecordBatch, ArrowError> {
     let schema = Arc::new(Schema::new(fields));
     let batch = RecordBatch::try_new(schema, columns)?;
     Ok(batch)
+}
+
+pub static COUNT_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|| {
+    Arc::new(Schema::new(vec![Field::new(
+        "count",
+        DataType::UInt64,
+        false,
+    )]))
+});
+
+pub fn create_count_record_batch(count: u64) -> RecordBatch {
+    RecordBatch::try_new(
+        COUNT_SCHEMA.clone(),
+        vec![Arc::new(UInt64Array::from_value(
+            count, /* rows = */ 1,
+        ))],
+    )
+    .unwrap()
 }
 
 #[cfg(test)]
